Relocate galaxy.datatypes.metadata to galaxy.model.metadata.
This eliminates the last dependency in galaxy.model on galaxy.datatypes - establishing a strict ordering of dependencies (galaxy.datatypes depends on galaxy.model and not vice versa).

The galaxy.datatypes.metadata interface needs to remain in place for tool shed datatypes and is perfectly reasonable to use from within the datatypes module - it now depends on galaxy.model.metadata and exports the same interfac.

In the longer run, this is a better, more explict design and should eliminate certain classes of problems with circular dependencies.

Fixes #1552
